requirements.txt
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
emoji==0.6.0
idna==2.10
mysql-connector-python==8.0.22
protobuf==3.14.0
pyTelegramBotAPI==3.7.3
python-dateutil==2.8.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.10